ai-demand-planner/requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
streamlit==1.35.0
python-dotenv==1.0.1